# product_docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmartFlow Agent Hub API",
    "description": "SmartFlow Agent Hub 接口文档，包含会话管理、知识库管理及 AI 对话核心接口。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
    "/conversations": {
      "get": {
        "tags": ["Conversations"],
        "summary": "获取会话列表",
        "description": "按更新时间倒序分页返回会话列表。",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户标识"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回会话列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Conversations"],
        "summary": "创建新会话",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id"],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "用户标识"
                  },
                  "title": {
                    "type": "string",
                    "description": "会话标题"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{session_id}": {
      "delete": {
        "tags": ["Conversations"],
        "summary": "删除会话",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户标识"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{session_id}/messages": {
      "get": {
        "tags": ["Chat"],
        "summary": "获取会话历史消息",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回历史消息",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/BaseResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Message"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/chat/completions": {
      "post": {
        "tags": ["Chat"],
        "summary": "发送消息 (Completions)",
        "description": "**Response (Stream - OpenAI Compatible):** \n遵循 Server-Sent Events (SSE) 标准，每行以 `data: ` 开头，结束发送 `data: [DONE]`。\n\n**1. 文本生成 (Standard)**\n```json\ndata: {\"id\":\"chatcmpl-123\",\"object\":\"chat.completion.chunk\",\"created\":1700000000,\"model\":\"smart-flow-agent-v1\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\"好\"},\"finish_reason\":null}]}\ndata: {\"id\":\"chatcmpl-123\",\"object\":\"chat.completion.chunk\",\"created\":1700000000,\"model\":\"smart-flow-agent-v1\",\"choices\":[{\"index\":0,\"delta\":{\"content\":\"的\"},\"finish_reason\":null}]}\n```\n\n**2. 思考过程 (Reasoning - DeepSeek/R1 Style)**\n使用 `reasoning_content` 字段透传 Agent 的规划与思考。\n```json\ndata: {\"id\":\"...\",\"choices\":[{\"index\":0,\"delta\":{\"reasoning_content\":\"正在检索知识库...\"},\"finish_reason\":null}]}\n```\n\n**3. 工具调用 (Tool Calls)**\n标准 OpenAI 工具调用格式。\n```json\ndata: {\"id\":\"...\",\"choices\":[{\"index\":0,\"delta\":{\"tool_calls\":[{\"index\":0,\"id\":\"call_abc\",\"type\":\"function\",\"function\":{\"name\":\"image_gen\",\"arguments\":\"\"}}]},\"finish_reason\":null}]}\ndata: {\"id\":\"...\",\"choices\":[{\"index\":0,\"delta\":{\"tool_calls\":[{\"index\":0,\"function\":{\"arguments\":\"{\\\"prompt\\\"\"}}]},\"finish_reason\":null}]}\n```\n\n**4. 扩展事件 (Extension Events)**\n为了支持 Agent 的特殊状态（如执行结果回显、审批请求），我们在 `choices[0]` 中添加 `extension` 字段。\n*注：标准 OpenAI SDK 会自动忽略此字段，不会报错；定制前端可解析此字段以增强体验。*\n\n*   **工具执行结果 (Tool Output)**:\n    ```json\n    {\n      \"choices\": [{\n        \"delta\": {},\n        \"extension\": {\n          \"type\": \"tool_result\",\n          \"tool_call_id\": \"call_abc\",\n          \"content\": \"https://image.url/gen.png\" // 图片或工具结果\n        }\n      }]\n    }\n    ```\n\n*   **请求审批 (Request Approval)**:\n    ```json\n    {\n      \"choices\": [{\n        \"finish_reason\": \"manual_approval_required\", // 自定义结束原因\n        \"extension\": {\n          \"type\": \"approval_request\",\n          \"approval_id\": \"appr_789\",\n          \"tool_call_id\": \"call_def\",\n          \"tool_name\": \"send_email\",\n          \"tool_args\": {\"to\": \"boss@company.com\"}\n        }\n      }]\n    }\n    ```\n\n**5. 结束 (Completion)**\n```json\ndata: [DONE]\n```",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应（支持流式或 JSON）",
            "content": {
              "text/event-stream": {
                "description": "SSE 流式响应"
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "标准 OpenAI 响应格式"
                }
              }
            }
          }
        }
      }
    },
    "/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "上传文档",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "用户标识"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "文件内容"
                  },
                  "session_id": {
                    "type": "string",
                    "description": "可选，若关联特定会话"
                  }
                },
                "required": [
                  "user_id",
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功上传文档",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "获取文档列表",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户标识"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "筛选状态"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回文档列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{doc_id}": {
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "删除文档",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "文档 ID"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户标识"
          }
        ],
        "responses": {
          "200": {
            "description": "成功删除文档",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BaseResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ConversationResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          {
            "type": "object",
            "properties": {
              "data": { "$ref": "#/components/schemas/Conversation" }
            }
          }
        ]
      },
      "ConversationListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Conversation" }
                  },
                  "total": { "type": "integer" },
                  "page": { "type": "integer" },
                  "size": { "type": "integer" }
                }
              }
            }
          }
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "type": { "type": "string", "enum": ["human", "ai"] },
          "content": { "type": "string" },
          "reasoning_content": { "type": "string", "nullable": true },
          "file_urls": {
            "type": "array",
            "items": { "type": "string" },
            "nullable": true
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["user_id", "model", "messages", "session_id"],
        "properties": {
          "user_id": { "type": "string", "description": "用户标识" },
          "model": { "type": "string", "example": "smart-flow-agent-v1" },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": { "type": "string", "enum": ["user", "assistant", "system"] },
                "content": { "type": "string" }
              }
            }
          },
          "stream": { "type": "boolean", "default": true },
          "session_id": { "type": "string", "description": "用于关联会话上下文" },
          "urls": {
            "type": "array",
            "items": { "type": "string" },
            "description": "本轮对话引用的文件链接列表"
          },
          "deep_thinking": {
            "type": "boolean",
            "default": false,
            "description": "是否开启深度思考"
          }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "filename": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "indexing",
              "indexed"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "DocumentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Document"
              }
            }
          }
        ]
      }
    }
  }
}